Add custom ordering of categories on dashboard
Those categories vary in significance and relevance so we want to order
them to highlight the most meaningful ones first. This is done with a
single 'order' attribute on a tag. They're presented in this order from
lowest number to highest.

Also changes the presentation of the categories on the dashboard so that
they collapse into a single list on smaller screen sizes.