Prevent horizontal scrolling on mobile
On mobile breakpoints the dashboard cover was causing the page to be
wider than the screen and enabling horizontal scrolling. The width of
the cover's content was set to 100vw ie 100% of the browser's width. As
this element has padding, the div ends up being 100vw + padding, so
wider than the screen.

This issue didn't show up on larger screens because the content
container also has the container class which sets a max-width of 100ch
ie 100 * the width of the 0 character.

Closes #145