Fix large spacing at end of scrolling-load
Pages with the load-as-you-scroll mechanic would end up with a large
amount of additional spacing. This spacing was caused by margins being
applied to hidden elements required for the loading.

As the loading depends on a few JS elements, we have included a
fallback. If the JS does not load or function correctly, the page will
show next and previous page links. These were hidden correctly, but the
.stack composition class was still applying margins.

The fallback links are now hidden by applying a .hidden class and this
class is specifically excluded from having margins applied in the .stack
composition class.

Closes #142