Add friendly URLs for entries and tags
Adds FriendlyId gem which manages a slug column. This will be populated
with a unique, url-safe string when an entry is created or updated (if a
slug does not exist). An extra method has been added to ensure that a
new slug is generated if the name is updated.

As FriendlyId can keep a history of changes to slugs to ensure old urls
don't break, tags#show and entries#show actions both redirect to the
latest slug.

Closes #122
Closes #123