Setup and document a docker-compose.yml file
Some users prefer running dependencies like Postgres in Docker. A
docker-compose.yml file makes it easy to ensure everyone is running the
right versions and to provide consistent configuration.

Also introduces .env.sample as a way to share non-sensitive environment
variables in the project. This just includes the `DATABASE_URL` for the
Postgres database if run on Docker.

ALso removes the reference to Redis as a dependency. This doesn't seem
to be used locally any more.

Closes #500